zuordnung/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
grafana-api==1.0.3
idna==2.10
influxdb==5.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
paho-mqtt==1.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
